fix: edit splitText directory name
