+++ src/utils/splitText/index.ts
