+++ src/utils/splitText/splitText.spec.ts
