+++ src/utils/splitText/splitText.ts
